Moved Clustering and RayCasting to the Algorithms namespace
