+++ Bot.Tests/RayCastingTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.Numerics;
+using Bot.Algorithms;
 using Bot.ExtensionMethods;
 
 namespace Bot.Tests;
