+++ Bot/ExtensionMethods/ArmyExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
+using Bot.Algorithms;
 using Bot.GameSense;
 using Bot.MapKnowledge;
