+++ Bot/ExtensionMethods/Vector2Extensions.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
+using Bot.Algorithms;
 using Bot.GameData;
 using Bot.GameSense.RegionTracking;
 using Bot.MapKnowledge;
