+++ Bot/ExtensionMethods/Vector3Extensions.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
+using Bot.Algorithms;
 using Bot.GameData;
 using Bot.MapKnowledge;
 using SC2APIProtocol;
