+++ Bot/MapKnowledge/RegionAnalyzer/ChokePoints/RayCastingChokeFinder.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
+using Bot.Algorithms;
 using Bot.Debugging.GraphicalDebugging;
 using Bot.ExtensionMethods;
 using Bot.Utils;
