+++ Bot/VideoClips/Clips/RayCastingClips/ChokeWallsClip.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
+using Bot.Algorithms;
 using Bot.ExtensionMethods;
 using Bot.MapKnowledge;
 using Bot.Utils;
@@ -17,7 +18,7 @@ public class ChokeWallsClip : Clip {
 
         var visibleCells = new HashSet<Vector2>();
         for (var angle = 0; angle < 360; angle += 1) {
-            var rayCastResults = Bot.RayCasting.RayCast(sceneLocation, MathUtils.DegToRad(angle), cell => !MapAnalyzer.IsWalkable(cell)).ToList();
+            var rayCastResults = RayCasting.RayCast(sceneLocation, MathUtils.DegToRad(angle), cell => !MapAnalyzer.IsWalkable(cell)).ToList();
             foreach (var rayCastResult in rayCastResults) {
                 visibleCells.Add(rayCastResult.CornerOfCell.AsWorldGridCenter());
             }
