+++ Bot/VideoClips/Clips/RayCastingClips/ChokeWidenessClip.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 using System.Numerics;
+using Bot.Algorithms;
 using Bot.Debugging.GraphicalDebugging;
 using Bot.ExtensionMethods;
 using Bot.MapKnowledge;
