+++ Bot/VideoClips/Clips/RayCastingClips/RayCastingIntersectionsClip.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using System.Numerics;
+using Bot.Algorithms;
 using Bot.ExtensionMethods;
 using Bot.MapKnowledge;
 using Bot.Utils;
