+++ Bot/VideoClips/Clips/RayCastingClips/RaySteppingClip.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
+using Bot.Algorithms;
 using Bot.ExtensionMethods;
 using Bot.MapKnowledge;
 using Bot.Utils;
