+++ Bot/VideoClips/Clips/RayCastingClips/SingleRayCastingClip.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 using System.Numerics;
+using Bot.Algorithms;
 using Bot.ExtensionMethods;
 using Bot.MapKnowledge;
 using Bot.Utils;
